% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_df.R
\name{create_df}
\alias{create_df}
\title{Create's a dataframe}
\usage{
create_df(n = 10)
}
\arguments{
\item{n}{number of rows}
}
\value{
data.frame
}
\description{
Create's a dataframe
}
\examples{
create_df()
}
